+++ deeplens/geolens.py
@@ -28,8 +28,8 @@ from torchvision.utils import make_grid, save_image
 from tqdm import tqdm
 from transformers import get_cosine_schedule_with_warmup
 
-from .lens import Lens
-from .optics.basics import (
+from deeplens.lens import Lens
+from deeplens.optics.basics import (
     DEFAULT_WAVE,
     DEPTH,
     EPSILON,
@@ -41,10 +41,10 @@ from .optics.basics import (
     SPP_RENDER,
     WAVE_RGB,
 )
-from .optics.materials import Material
-from .optics.monte_carlo import forward_integral
-from .optics.ray import Ray
-from .optics.geometric_surface import (
+from deeplens.optics.materials import Material
+from deeplens.optics.monte_carlo import forward_integral
+from deeplens.optics.ray import Ray
+from deeplens.optics.geometric_surface import (
     Aperture,
     Aspheric,
     Cubic,
@@ -53,9 +53,9 @@ from .optics.geometric_surface import (
     Spheric,
     ThinLens,
 )
-from .optics.wave import AngularSpectrumMethod
-from .optics.utils import diff_float
-from .utils import (
+from deeplens.optics.wave import AngularSpectrumMethod
+from deeplens.optics.utils import diff_float
+from deeplens.utils import (
     batch_psnr,
     batch_ssim,
     denormalize_ImageNet,
@@ -63,10 +63,10 @@ from .utils import (
     normalize_ImageNet,
     set_logger,
 )
-# from .geolens_utils import draw_lens_layout, draw_layout_3d
+from deeplens.geolens_eval import GeoLensEval
+from deeplens.geolens_optim import GeoLensOptim
 
-
-class GeoLens(Lens):
+class GeoLens(Lens, GeoLensEval, GeoLensOptim):
     def __init__(self, filename=None, sensor_res=(1000, 1000), sensor_size=(8.0, 8.0), device=None):
         """Initialize a refractive lens.
         
@@ -244,7 +244,7 @@ class GeoLens(Lens):
         ray_o = self.sample_circle(pupilr, pupilz, shape=[len(fov_x), len(fov_y), num_rays])
 
         # Sample ray directions, shape [num_fov_x, num_fov_y, num_rays, 3]
-        fx_grid, fy_grid = torch.meshgrid(fov_x, -fov_y, indexing="xy")
+        fx_grid, fy_grid = torch.meshgrid(fov_x, fov_y, indexing="ij")
         dx = torch.tan(fx_grid).unsqueeze(-1).expand_as(ray_o[..., 0])
         dy = torch.tan(fy_grid).unsqueeze(-1).expand_as(ray_o[..., 1])
         dz = torch.ones_like(ray_o[..., 2])
@@ -1213,42 +1213,42 @@ class GeoLens(Lens):
     # ====================================================================================
     # Classical optical design
     # ====================================================================================
-    @torch.no_grad()
-    def rms_map(self, res=(128, 128), depth=DEPTH):
-        """Calculate the RMS spot error map as a weight mask for lens design.
+    # @torch.no_grad()
+    # def rms_map(self, res=(128, 128), depth=DEPTH):
+    #     """Calculate the RMS spot error map as a weight mask for lens design.
 
-        Args:
-            res (tuple, optional): resolution of the RMS map. Defaults to (32, 32).
-            depth (float, optional): depth of the point source. Defaults to DEPTH.
+    #     Args:
+    #         res (tuple, optional): resolution of the RMS map. Defaults to (32, 32).
+    #         depth (float, optional): depth of the point source. Defaults to DEPTH.
 
-        Returns:
-            rms_map (torch.Tensor): RMS map normalized to [0, 1].
-        """
-        ray = self.sample_point_source(depth=depth, num_rays=SPP_PSF, num_grid=64)
-        ray, _ = self.trace(ray)
-        o2 = ray.project_to(self.d_sensor)
-        o2_center = (o2 * ray.ra.unsqueeze(-1)).sum(0) / ray.ra.sum(0).add(
-            EPSILON
-        ).unsqueeze(-1)
-        # normalized to center (0, 0)
-        o2_norm = (o2 - o2_center) * ray.ra.unsqueeze(-1)
+    #     Returns:
+    #         rms_map (torch.Tensor): RMS map normalized to [0, 1].
+    #     """
+    #     ray = self.sample_point_source(depth=depth, num_rays=SPP_PSF, num_grid=64)
+    #     ray, _ = self.trace(ray)
+    #     o2 = ray.project_to(self.d_sensor)
+    #     o2_center = (o2 * ray.ra.unsqueeze(-1)).sum(0) / ray.ra.sum(0).add(
+    #         EPSILON
+    #     ).unsqueeze(-1)
+    #     # normalized to center (0, 0)
+    #     o2_norm = (o2 - o2_center) * ray.ra.unsqueeze(-1)
 
-        rms_map = torch.sqrt(
-            ((o2_norm**2).sum(-1) * ray.ra).sum(0) / (ray.ra.sum(0) + EPSILON)
-        )
-        rms_map = (
-            F.interpolate(
-                rms_map.unsqueeze(0).unsqueeze(0),
-                res,
-                mode="bilinear",
-                align_corners=True,
-            )
-            .squeeze(0)
-            .squeeze(0)
-        )
-        rms_map /= rms_map.max()
+    #     rms_map = torch.sqrt(
+    #         ((o2_norm**2).sum(-1) * ray.ra).sum(0) / (ray.ra.sum(0) + EPSILON)
+    #     )
+    #     rms_map = (
+    #         F.interpolate(
+    #             rms_map.unsqueeze(0).unsqueeze(0),
+    #             res,
+    #             mode="bilinear",
+    #             align_corners=True,
+    #         )
+    #         .squeeze(0)
+    #         .squeeze(0)
+    #     )
+    #     rms_map /= rms_map.max()
 
-        return rms_map
+    #     return rms_map
 
     def analysis_rms(self, depth=float("inf")):
         """Compute RMS-based error. Contain both RMS errors and RMS radius. This function needs more testing."""
@@ -1357,145 +1357,145 @@ class GeoLens(Lens):
                 f"RMS average error (chief ray): center {rms_radius_on_axis} mm, off-axis {rms_radius_off_axis} mm, average {rms_avg} mm"
             )
 
-    def psf2mtf(self, psf, diag=False):
-        """Convert 2D PSF kernel to MTF curve by FFT.
+    # def psf2mtf(self, psf, diag=False):
+    #     """Convert 2D PSF kernel to MTF curve by FFT.
 
-        Args:
-            psf (tensor): 2D PSF tensor.
+    #     Args:
+    #         psf (tensor): 2D PSF tensor.
 
-        Returns:
-            freq (ndarray): Frequency axis.
-            tangential_mtf (ndarray): Tangential MTF.
-            sagittal_mtf (ndarray): Sagittal MTF.
-        """
-        psf = psf.cpu().numpy()
-        x = np.linspace(-1, 1, psf.shape[1]) * self.pixel_size * psf.shape[1] / 2
-        y = np.linspace(-1, 1, psf.shape[0]) * self.pixel_size * psf.shape[0] / 2
+    #     Returns:
+    #         freq (ndarray): Frequency axis.
+    #         tangential_mtf (ndarray): Tangential MTF.
+    #         sagittal_mtf (ndarray): Sagittal MTF.
+    #     """
+    #     psf = psf.cpu().numpy()
+    #     x = np.linspace(-1, 1, psf.shape[1]) * self.pixel_size * psf.shape[1] / 2
+    #     y = np.linspace(-1, 1, psf.shape[0]) * self.pixel_size * psf.shape[0] / 2
 
-        if diag:
-            raise Exception("Diagonal PSF is not tested.")
-            diag_psf = np.diag(np.flip(psf, axis=0))
-            x *= math.sqrt(2)
-            y *= math.sqrt(2)
-            delta_x = self.pixel_size * math.sqrt(2)
+    #     if diag:
+    #         raise Exception("Diagonal PSF is not tested.")
+    #         diag_psf = np.diag(np.flip(psf, axis=0))
+    #         x *= math.sqrt(2)
+    #         y *= math.sqrt(2)
+    #         delta_x = self.pixel_size * math.sqrt(2)
 
-            diag_mtf = np.abs(np.fft.fft(diag_psf))
-            # diag_mtf /= diag_mtf.max()
+    #         diag_mtf = np.abs(np.fft.fft(diag_psf))
+    #         # diag_mtf /= diag_mtf.max()
 
-            # Create frequency axis in cycles/mm
-            freq = np.fft.fftfreq(psf.shape[0], delta_x)
+    #         # Create frequency axis in cycles/mm
+    #         freq = np.fft.fftfreq(psf.shape[0], delta_x)
 
-            # Only keep the positive frequencies
-            positive_freq_idx = freq > 0
+    #         # Only keep the positive frequencies
+    #         positive_freq_idx = freq > 0
 
-            freq = freq[positive_freq_idx]
-            diag_mtf = diag_mtf[positive_freq_idx]
-            diag_mtf /= diag_mtf[0]
+    #         freq = freq[positive_freq_idx]
+    #         diag_mtf = diag_mtf[positive_freq_idx]
+    #         diag_mtf /= diag_mtf[0]
 
-            return freq, diag_mtf
-        else:
-            # Extract 1D PSFs along the sagittal and tangential directions
-            center_x = psf.shape[1] // 2
-            center_y = psf.shape[0] // 2
-            sagittal_psf = psf[center_y, :]
-            tangential_psf = psf[:, center_x]
+    #         return freq, diag_mtf
+    #     else:
+    #         # Extract 1D PSFs along the sagittal and tangential directions
+    #         center_x = psf.shape[1] // 2
+    #         center_y = psf.shape[0] // 2
+    #         sagittal_psf = psf[center_y, :]
+    #         tangential_psf = psf[:, center_x]
 
-            # Fourier Transform to get the MTFs
-            sagittal_mtf = np.abs(np.fft.fft(sagittal_psf))
-            tangential_mtf = np.abs(np.fft.fft(tangential_psf))
+    #         # Fourier Transform to get the MTFs
+    #         sagittal_mtf = np.abs(np.fft.fft(sagittal_psf))
+    #         tangential_mtf = np.abs(np.fft.fft(tangential_psf))
 
-            # Normalize the MTFs
-            sagittal_mtf /= sagittal_mtf.max()
-            tangential_mtf /= tangential_mtf.max()
+    #         # Normalize the MTFs
+    #         sagittal_mtf /= sagittal_mtf.max()
+    #         tangential_mtf /= tangential_mtf.max()
 
-            delta_x = self.pixel_size  # / 2
+    #         delta_x = self.pixel_size  # / 2
 
-            # Create frequency axis in cycles/mm
-            freq = np.fft.fftfreq(psf.shape[0], delta_x)
+    #         # Create frequency axis in cycles/mm
+    #         freq = np.fft.fftfreq(psf.shape[0], delta_x)
 
-            # Only keep the positive frequencies
-            positive_freq_idx = freq > 0
+    #         # Only keep the positive frequencies
+    #         positive_freq_idx = freq > 0
 
-            return (
-                freq[positive_freq_idx],
-                tangential_mtf[positive_freq_idx],
-                sagittal_mtf[positive_freq_idx],
-            )
+    #         return (
+    #             freq[positive_freq_idx],
+    #             tangential_mtf[positive_freq_idx],
+    #             sagittal_mtf[positive_freq_idx],
+    #         )
 
-    def vignetting(self):
-        """Compute vignetting."""
-        pass
+    # def vignetting(self):
+    #     """Compute vignetting."""
+    #     pass
 
-    def field_curvature(self):
-        """Compute field curvature."""
-        pass
+    # def field_curvature(self):
+    #     """Compute field curvature."""
+    #     pass
 
-    def ray_aberration(self):
-        """Compute ray aberration."""
-        pass
+    # def ray_aberration(self):
+    #     """Compute ray aberration."""
+    #     pass
 
-    def distortion(self, depth=DEPTH, grid_size=64):
-        """Compute distortion map at a given depth.
+    # def distortion(self, depth=DEPTH, grid_size=64):
+    #     """Compute distortion map at a given depth.
 
-        Args:
-            depth (float): depth of the point source.
-            img_res (tuple): resolution of the image.
+    #     Args:
+    #         depth (float): depth of the point source.
+    #         img_res (tuple): resolution of the image.
 
-        Returns:
-            distortion_grid (torch.Tensor): distortion map. shape (grid_size, grid_size, 2)
-        """
-        # Ray tracing to calculate distortion map
-        ray = self.sample_point_source(
-            depth=depth, num_rays=SPP_CALC, num_grid=grid_size
-        )
-        ray = self.trace2sensor(ray)
-        o_dist = (ray.o[..., :2] * ray.ra.unsqueeze(-1)).sum(-2) / ray.ra.unsqueeze(
-            -1
-        ).sum(-2).add(EPSILON)  # shape (H, W, 2)
+    #     Returns:
+    #         distortion_grid (torch.Tensor): distortion map. shape (grid_size, grid_size, 2)
+    #     """
+    #     # Ray tracing to calculate distortion map
+    #     ray = self.sample_point_source(
+    #         depth=depth, num_rays=SPP_CALC, num_grid=grid_size
+    #     )
+    #     ray = self.trace2sensor(ray)
+    #     o_dist = (ray.o[..., :2] * ray.ra.unsqueeze(-1)).sum(-2) / ray.ra.unsqueeze(
+    #         -1
+    #     ).sum(-2).add(EPSILON)  # shape (H, W, 2)
 
-        x_dist = -o_dist[..., 0] / self.sensor_size[1] * 2
-        y_dist = o_dist[..., 1] / self.sensor_size[0] * 2
-        distortion_grid = torch.stack((x_dist, y_dist), dim=-1)  # shape (H, W, 2)
-        return distortion_grid
+    #     x_dist = -o_dist[..., 0] / self.sensor_size[1] * 2
+    #     y_dist = o_dist[..., 1] / self.sensor_size[0] * 2
+    #     distortion_grid = torch.stack((x_dist, y_dist), dim=-1)  # shape (H, W, 2)
+    #     return distortion_grid
 
-    def calc_distortion_1D(self, hfov, wvln=DEFAULT_WAVE, plane="meridional", ray_aiming=True):
-        """Calculate distortion at a specific field angle.
+    # def calc_distortion_1D(self, hfov, wvln=DEFAULT_WAVE, plane="meridional", ray_aiming=True):
+    #     """Calculate distortion at a specific field angle.
 
-        Args:
-            hfov (float): view angle (degree)
-            plane (str): meridional or sagittal
-            ray_aiming (bool): whether the chief ray through the center of the stop.
+    #     Args:
+    #         hfov (float): view angle (degree)
+    #         plane (str): meridional or sagittal
+    #         ray_aiming (bool): whether the chief ray through the center of the stop.
 
-        Returns:
-            distortion (float): distortion at the specific field angle
-        """
-        # Calculate ideal image height
-        effective_foclen = self.calc_efl()
-        ideal_image_height = effective_foclen * torch.tan(hfov * torch.pi / 180)
+    #     Returns:
+    #         distortion (float): distortion at the specific field angle
+    #     """
+    #     # Calculate ideal image height
+    #     effective_foclen = self.calc_efl()
+    #     ideal_image_height = effective_foclen * torch.tan(hfov * torch.pi / 180)
 
-        # Calculate chief ray
-        chief_ray_o, chief_ray_d = self.calc_chief_ray_infinite(hfov=hfov, wvln=wvln, plane=plane, ray_aiming=ray_aiming)
-        ray = Ray(chief_ray_o, chief_ray_d, wvln=wvln, device=self.device)
+    #     # Calculate chief ray
+    #     chief_ray_o, chief_ray_d = self.calc_chief_ray_infinite(hfov=hfov, wvln=wvln, plane=plane, ray_aiming=ray_aiming)
+    #     ray = Ray(chief_ray_o, chief_ray_d, wvln=wvln, device=self.device)
 
-        ray, _ = self.trace(ray, lens_range=range(len(self.surfaces)))
-        t = (self.d_sensor - ray.o[..., 2]) / ray.d[..., 2]
+    #     ray, _ = self.trace(ray, lens_range=range(len(self.surfaces)))
+    #     t = (self.d_sensor - ray.o[..., 2]) / ray.d[..., 2]
 
-        # Calculate actual image height
-        if plane == "sagittal":
-            actual_image_height = abs(ray.o[..., 0] + ray.d[..., 0] * t)
-        elif plane == "meridional":
-            actual_image_height = abs(ray.o[..., 1] + ray.d[..., 1] * t)
-        else:
-            raise ValueError(f"Invalid plane: {plane}")
+    #     # Calculate actual image height
+    #     if plane == "sagittal":
+    #         actual_image_height = abs(ray.o[..., 0] + ray.d[..., 0] * t)
+    #     elif plane == "meridional":
+    #         actual_image_height = abs(ray.o[..., 1] + ray.d[..., 1] * t)
+    #     else:
+    #         raise ValueError(f"Invalid plane: {plane}")
 
-        # Calculate distortion
-        distortion = (actual_image_height - ideal_image_height) / ideal_image_height
+    #     # Calculate distortion
+    #     distortion = (actual_image_height - ideal_image_height) / ideal_image_height
 
-        # Handle the case where ideal_image_height is 0 or very close to 0
-        mask = abs(ideal_image_height) < EPSILON
-        distortion[mask] = torch.tensor(0.0, device=self.device)
+    #     # Handle the case where ideal_image_height is 0 or very close to 0
+    #     mask = abs(ideal_image_height) < EPSILON
+    #     distortion[mask] = torch.tensor(0.0, device=self.device)
 
-        return distortion
+    #     return distortion
 
     # ====================================================================================
     # Geometrical optics calculation
@@ -2494,627 +2494,627 @@ class GeoLens(Lens):
             lens_title=lens_title,
         )
 
-    def draw_layout_3d(self, filename=None, figsize=(10, 6), view_angle=30, show=True):
-        """Draw 3D layout of the lens system.
+    # def draw_layout_3d(self, filename=None, figsize=(10, 6), view_angle=30, show=True):
+    #     """Draw 3D layout of the lens system.
 
-        Args:
-            filename (str, optional): Path to save the figure. Defaults to None.
-            figsize (tuple): Figure size
-            view_angle (int): Viewing angle for the 3D plot
-            show (bool): Whether to display the figure
+    #     Args:
+    #         filename (str, optional): Path to save the figure. Defaults to None.
+    #         figsize (tuple): Figure size
+    #         view_angle (int): Viewing angle for the 3D plot
+    #         show (bool): Whether to display the figure
 
-        Returns:
-            fig, ax: Matplotlib figure and axis objects
-        """
-        from deeplens.geolens_utils import draw_layout_3d
-        return draw_layout_3d(
-            self, filename=filename, figsize=figsize, view_angle=view_angle, show=show
-        )
+    #     Returns:
+    #         fig, ax: Matplotlib figure and axis objects
+    #     """
+    #     from deeplens.geolens_utils import draw_layout_3d
+    #     return draw_layout_3d(
+    #         self, filename=filename, figsize=figsize, view_angle=view_angle, show=show
+    #     )
 
-    @torch.no_grad()
-    def draw_psf_radial(
-        self, M=3, depth=DEPTH, ks=PSF_KS, log_scale=False, save_name="./psf_radial.png"
-    ):
-        """Draw radial PSF (45 deg). Will draw M PSFs, each of size ks x ks."""
-        x = torch.linspace(0, 1, M)
-        y = torch.linspace(0, 1, M)
-        z = torch.full_like(x, depth)
-        points = torch.stack((x, y, z), dim=-1)
+    # @torch.no_grad()
+    # def draw_psf_radial(
+    #     self, M=3, depth=DEPTH, ks=PSF_KS, log_scale=False, save_name="./psf_radial.png"
+    # ):
+    #     """Draw radial PSF (45 deg). Will draw M PSFs, each of size ks x ks."""
+    #     x = torch.linspace(0, 1, M)
+    #     y = torch.linspace(0, 1, M)
+    #     z = torch.full_like(x, depth)
+    #     points = torch.stack((x, y, z), dim=-1)
 
-        psfs = []
-        for i in range(M):
-            # Scale PSF for a better visualization
-            psf = self.psf_rgb(points=points[i], ks=ks, center=True, spp=4096)
-            psf /= psf.max()
+    #     psfs = []
+    #     for i in range(M):
+    #         # Scale PSF for a better visualization
+    #         psf = self.psf_rgb(points=points[i], ks=ks, center=True, spp=4096)
+    #         psf /= psf.max()
 
-            if log_scale:
-                psf = torch.log(psf + EPSILON)
-                psf = (psf - psf.min()) / (psf.max() - psf.min())
+    #         if log_scale:
+    #             psf = torch.log(psf + EPSILON)
+    #             psf = (psf - psf.min()) / (psf.max() - psf.min())
 
-            psfs.append(psf)
+    #         psfs.append(psf)
 
-        psf_grid = make_grid(psfs, nrow=M, padding=1, pad_value=0.0)
-        save_image(psf_grid, save_name, normalize=True)
+    #     psf_grid = make_grid(psfs, nrow=M, padding=1, pad_value=0.0)
+    #     save_image(psf_grid, save_name, normalize=True)
 
-    @torch.no_grad()
-    def draw_spot_diagram(self, M=5, depth=DEPTH, wvln=DEFAULT_WAVE, save_name=None):
-        """Draw spot diagram of the lens. Shot rays from grid points in object space, trace to sensor and visualize."""
-        # Sample and trace rays from grid points
-        ray = self.sample_point_source(
-            depth=depth,
-            num_rays=SPP_CALC,
-            num_grid=GEO_GRID,
-            wvln=wvln,
-        )
-        ray = self.trace2sensor(ray)
-        o2 = -ray.o.clone().cpu().numpy()
-        ra = ray.ra.clone().cpu().numpy()
+    # @torch.no_grad()
+    # def draw_spot_diagram(self, M=5, depth=DEPTH, wvln=DEFAULT_WAVE, save_name=None):
+    #     """Draw spot diagram of the lens. Shot rays from grid points in object space, trace to sensor and visualize."""
+    #     # Sample and trace rays from grid points
+    #     ray = self.sample_point_source(
+    #         depth=depth,
+    #         num_rays=SPP_CALC,
+    #         num_grid=GEO_GRID,
+    #         wvln=wvln,
+    #     )
+    #     ray = self.trace2sensor(ray)
+    #     o2 = -ray.o.clone().cpu().numpy()
+    #     ra = ray.ra.clone().cpu().numpy()
 
-        # Plot multiple spot diagrams in one figure
-        fig, axs = plt.subplots(M, M, figsize=(30, 30))
-        for i in range(M):
-            for j in range(M):
-                ra_ = ra[:, i, j]
-                x, y = o2[:, i, j, 0], o2[:, i, j, 1]
-                x, y = x[ra_ > 0], y[ra_ > 0]
-                xc, yc = x.sum() / ra_.sum(), y.sum() / ra_.sum()
+    #     # Plot multiple spot diagrams in one figure
+    #     fig, axs = plt.subplots(M, M, figsize=(30, 30))
+    #     for i in range(M):
+    #         for j in range(M):
+    #             ra_ = ra[:, i, j]
+    #             x, y = o2[:, i, j, 0], o2[:, i, j, 1]
+    #             x, y = x[ra_ > 0], y[ra_ > 0]
+    #             xc, yc = x.sum() / ra_.sum(), y.sum() / ra_.sum()
 
-                # scatter plot
-                axs[i, j].scatter(x, y, 1, "black")
-                axs[i, j].scatter([xc], [yc], None, "r", "x")
-                axs[i, j].set_aspect("equal", adjustable="datalim")
+    #             # scatter plot
+    #             axs[i, j].scatter(x, y, 1, "black")
+    #             axs[i, j].scatter([xc], [yc], None, "r", "x")
+    #             axs[i, j].set_aspect("equal", adjustable="datalim")
 
-        if save_name is None:
-            save_name = f"./spot{-depth}mm.png"
-        else:
-            save_name = f"{save_name}_spot{-depth}mm.png"
+    #     if save_name is None:
+    #         save_name = f"./spot{-depth}mm.png"
+    #     else:
+    #         save_name = f"{save_name}_spot{-depth}mm.png"
 
-        plt.savefig(save_name, bbox_inches="tight", format="png", dpi=300)
-        plt.close()
+    #     plt.savefig(save_name, bbox_inches="tight", format="png", dpi=300)
+    #     plt.close()
 
-    @torch.no_grad()
-    def draw_spot_radial(self, M=3, depth=DEPTH, save_name=None):
-        """Draw radial spot diagram of the lens.
+    # @torch.no_grad()
+    # def draw_spot_radial(self, M=3, depth=DEPTH, save_name=None):
+    #     """Draw radial spot diagram of the lens.
 
-        Args:
-            M (int, optional): field number. Defaults to 3.
-            depth (float, optional): depth of the point source. Defaults to DEPTH.
-            save_name (string, optional): filename to save. Defaults to None.
-        """
-        # Sample and trace rays
-        # mag = self.calc_magnification3(depth)
-        ray = self.sample_point_source(
-            depth=depth,
-            num_rays=1024,
-            num_grid=M * 2 - 1,
-            wvln=DEFAULT_WAVE,
-        )
-        ray, _ = self.trace(ray)
-        ray.propagate_to(self.d_sensor)
-        o2 = torch.flip(ray.o.clone(), [1, 2]).cpu().numpy()
-        ra = torch.flip(ray.ra.clone(), [1, 2]).cpu().numpy()
+    #     Args:
+    #         M (int, optional): field number. Defaults to 3.
+    #         depth (float, optional): depth of the point source. Defaults to DEPTH.
+    #         save_name (string, optional): filename to save. Defaults to None.
+    #     """
+    #     # Sample and trace rays
+    #     # mag = self.calc_magnification3(depth)
+    #     ray = self.sample_point_source(
+    #         depth=depth,
+    #         num_rays=1024,
+    #         num_grid=M * 2 - 1,
+    #         wvln=DEFAULT_WAVE,
+    #     )
+    #     ray, _ = self.trace(ray)
+    #     ray.propagate_to(self.d_sensor)
+    #     o2 = torch.flip(ray.o.clone(), [1, 2]).cpu().numpy()
+    #     ra = torch.flip(ray.ra.clone(), [1, 2]).cpu().numpy()
 
-        # Plot multiple spot diagrams in one figure
-        fig, axs = plt.subplots(1, M, figsize=(M * 12, 10))
-        for i in range(M):
-            i_bias = i + M - 1
+    #     # Plot multiple spot diagrams in one figure
+    #     fig, axs = plt.subplots(1, M, figsize=(M * 12, 10))
+    #     for i in range(M):
+    #         i_bias = i + M - 1
 
-            # calculate center of mass
-            ra_ = ra[:, i_bias, i_bias]
-            x, y = o2[:, i_bias, i_bias, 0], o2[:, i_bias, i_bias, 1]
-            x, y = x[ra_ > 0], y[ra_ > 0]
-            xc, yc = x.sum() / ra_.sum(), y.sum() / ra_.sum()
+    #         # calculate center of mass
+    #         ra_ = ra[:, i_bias, i_bias]
+    #         x, y = o2[:, i_bias, i_bias, 0], o2[:, i_bias, i_bias, 1]
+    #         x, y = x[ra_ > 0], y[ra_ > 0]
+    #         xc, yc = x.sum() / ra_.sum(), y.sum() / ra_.sum()
 
-            # scatter plot
-            axs[i].scatter(x, y, 12, "black")
-            axs[i].scatter([xc], [yc], 400, "r", "x")
+    #         # scatter plot
+    #         axs[i].scatter(x, y, 12, "black")
+    #         axs[i].scatter([xc], [yc], 400, "r", "x")
 
-            # visualization
-            axs[i].set_aspect("equal", adjustable="datalim")
-            axs[i].tick_params(axis="both", which="major", labelsize=18)
-            axs[i].spines["top"].set_linewidth(4)
-            axs[i].spines["bottom"].set_linewidth(4)
-            axs[i].spines["left"].set_linewidth(4)
-            axs[i].spines["right"].set_linewidth(4)
+    #         # visualization
+    #         axs[i].set_aspect("equal", adjustable="datalim")
+    #         axs[i].tick_params(axis="both", which="major", labelsize=18)
+    #         axs[i].spines["top"].set_linewidth(4)
+    #         axs[i].spines["bottom"].set_linewidth(4)
+    #         axs[i].spines["left"].set_linewidth(4)
+    #         axs[i].spines["right"].set_linewidth(4)
 
-        # Save figure
-        if save_name is None:
-            plt.savefig(
-                f"./spot{-depth}mm_radial.svg",
-                bbox_inches="tight",
-                format="svg",
-                dpi=1200,
-            )
-        else:
-            plt.savefig(
-                f"{save_name}_spot{-depth}mm_radial.svg",
-                bbox_inches="tight",
-                format="svg",
-                dpi=1200,
-            )
+    #     # Save figure
+    #     if save_name is None:
+    #         plt.savefig(
+    #             f"./spot{-depth}mm_radial.svg",
+    #             bbox_inches="tight",
+    #             format="svg",
+    #             dpi=1200,
+    #         )
+    #     else:
+    #         plt.savefig(
+    #             f"{save_name}_spot{-depth}mm_radial.svg",
+    #             bbox_inches="tight",
+    #             format="svg",
+    #             dpi=1200,
+    #         )
 
-        plt.close()
+    #     plt.close()
 
-    @torch.no_grad()
-    def draw_mtf(
-        self,
-        wvlns=DEFAULT_WAVE,
-        depth=DEPTH,
-        relative_fov=[0.0, 0.7, 1.0],
-        save_name="./mtf.png",
-    ):
-        """Draw MTF curve (different FoVs, single wvln, infinite depth) of the lens."""
-        assert save_name[-4:] == ".png", "save_name must end with .png"
+    # @torch.no_grad()
+    # def draw_mtf(
+    #     self,
+    #     wvlns=DEFAULT_WAVE,
+    #     depth=DEPTH,
+    #     relative_fov=[0.0, 0.7, 1.0],
+    #     save_name="./mtf.png",
+    # ):
+    #     """Draw MTF curve (different FoVs, single wvln, infinite depth) of the lens."""
+    #     assert save_name[-4:] == ".png", "save_name must end with .png"
 
-        relative_fov = (
-            [relative_fov] if isinstance(relative_fov, float) else relative_fov
-        )
-        wvlns = [wvlns] if isinstance(wvlns, float) else wvlns
-        color_list = "rgb"
+    #     relative_fov = (
+    #         [relative_fov] if isinstance(relative_fov, float) else relative_fov
+    #     )
+    #     wvlns = [wvlns] if isinstance(wvlns, float) else wvlns
+    #     color_list = "rgb"
 
-        plt.figure(figsize=(6, 6))
-        for wvln_idx, wvln in enumerate(wvlns):
-            for fov_idx, fov in enumerate(relative_fov):
-                point = torch.tensor([0, fov, depth])
-                psf = self.psf(points=point, wvln=wvln, ks=256)
-                freq, mtf_tan, mtf_sag = self.psf2mtf(psf)
+    #     plt.figure(figsize=(6, 6))
+    #     for wvln_idx, wvln in enumerate(wvlns):
+    #         for fov_idx, fov in enumerate(relative_fov):
+    #             point = torch.tensor([0, fov, depth])
+    #             psf = self.psf(points=point, wvln=wvln, ks=256)
+    #             freq, mtf_tan, mtf_sag = self.psf2mtf(psf)
 
-                fov_deg = round(fov * self.hfov * 180 / torch.pi, 1)
-                plt.plot(
-                    freq,
-                    mtf_tan,
-                    color_list[fov_idx],
-                    label=f"{fov_deg}(deg)-Tangential",
-                )
-                plt.plot(
-                    freq,
-                    mtf_sag,
-                    color_list[fov_idx],
-                    label=f"{fov_deg}(deg)-Sagittal",
-                    linestyle="--",
-                )
+    #             fov_deg = round(fov * self.hfov * 180 / torch.pi, 1)
+    #             plt.plot(
+    #                 freq,
+    #                 mtf_tan,
+    #                 color_list[fov_idx],
+    #                 label=f"{fov_deg}(deg)-Tangential",
+    #             )
+    #             plt.plot(
+    #                 freq,
+    #                 mtf_sag,
+    #                 color_list[fov_idx],
+    #                 label=f"{fov_deg}(deg)-Sagittal",
+    #                 linestyle="--",
+    #             )
 
-        plt.legend()
-        plt.xlabel("Spatial Frequency [cycles/mm]")
-        plt.ylabel("MTF")
+    #     plt.legend()
+    #     plt.xlabel("Spatial Frequency [cycles/mm]")
+    #     plt.ylabel("MTF")
 
-        # Save figure
-        plt.savefig(f"{save_name}", bbox_inches="tight", format="png", dpi=300)
-        plt.close()
+    #     # Save figure
+    #     plt.savefig(f"{save_name}", bbox_inches="tight", format="png", dpi=300)
+    #     plt.close()
 
-    def draw_distortion(self, filename=None, depth=DEPTH, grid_size=16):
-        """Draw distortion."""
-        # Ray tracing to calculate distortion map
-        distortion_grid = self.distortion(depth=depth, grid_size=grid_size)
-        x1 = distortion_grid[..., 0].cpu().numpy()
-        y1 = distortion_grid[..., 1].cpu().numpy()
+    # def draw_distortion(self, filename=None, depth=DEPTH, grid_size=16):
+    #     """Draw distortion."""
+    #     # Ray tracing to calculate distortion map
+    #     distortion_grid = self.distortion(depth=depth, grid_size=grid_size)
+    #     x1 = distortion_grid[..., 0].cpu().numpy()
+    #     y1 = distortion_grid[..., 1].cpu().numpy()
 
-        # Draw image
-        fig, ax = plt.subplots()
-        ax.set_title("Lens distortion")
-        ax.scatter(x1, y1, s=2)
-        ax.axis("scaled")
-        ax.grid(True)
+    #     # Draw image
+    #     fig, ax = plt.subplots()
+    #     ax.set_title("Lens distortion")
+    #     ax.scatter(x1, y1, s=2)
+    #     ax.axis("scaled")
+    #     ax.grid(True)
 
-        # Add grid lines based on grid_size
-        ax.set_xticks(np.linspace(-1, 1, grid_size))
-        ax.set_yticks(np.linspace(-1, 1, grid_size))
+    #     # Add grid lines based on grid_size
+    #     ax.set_xticks(np.linspace(-1, 1, grid_size))
+    #     ax.set_yticks(np.linspace(-1, 1, grid_size))
 
-        if filename is None:
-            plt.savefig(
-                f"./distortion{-depth}mm.png",
-                bbox_inches="tight",
-                format="png",
-                dpi=300,
-            )
-        else:
-            plt.savefig(
-                f"{filename[:-4]}_distortion_{-depth}mm.png",
-                bbox_inches="tight",
-                format="png",
-                dpi=300,
-            )
+    #     if filename is None:
+    #         plt.savefig(
+    #             f"./distortion{-depth}mm.png",
+    #             bbox_inches="tight",
+    #             format="png",
+    #             dpi=300,
+    #         )
+    #     else:
+    #         plt.savefig(
+    #             f"{filename[:-4]}_distortion_{-depth}mm.png",
+    #             bbox_inches="tight",
+    #             format="png",
+    #             dpi=300,
+    #         )
 
-    def draw_distortion_1D(
-        self, hfov, filename=None, num_points=GEO_GRID, wvln=DEFAULT_WAVE, plane="meridional", ray_aiming=True
-    ):
-        """Draw distortion. zemax format(default): ray_aiming = False.
+    # def draw_distortion_1D(
+    #     self, hfov, filename=None, num_points=GEO_GRID, wvln=DEFAULT_WAVE, plane="meridional", ray_aiming=True
+    # ):
+    #     """Draw distortion. zemax format(default): ray_aiming = False.
 
-        Args:
-            hfov: view angle (degrees)
-            filename: Save filename. Defaults to None.
-            num_points: Number of points. Defaults to GEO_GRID.
-            plane: Meridional or sagittal. Defaults to meridional.
-            ray_aiming: Whether to use ray aiming. Defaults to False.
-        """
+    #     Args:
+    #         hfov: view angle (degrees)
+    #         filename: Save filename. Defaults to None.
+    #         num_points: Number of points. Defaults to GEO_GRID.
+    #         plane: Meridional or sagittal. Defaults to meridional.
+    #         ray_aiming: Whether to use ray aiming. Defaults to False.
+    #     """
 
-        # Sample view angles
-        hfov_samples = torch.linspace(0, hfov, num_points)
-        distortions = []
+    #     # Sample view angles
+    #     hfov_samples = torch.linspace(0, hfov, num_points)
+    #     distortions = []
 
-        # Calculate distortion
-        distortions = self.calc_distortion_1D(hfov=hfov_samples, wvln=wvln, plane=plane, ray_aiming=ray_aiming)
+    #     # Calculate distortion
+    #     distortions = self.calc_distortion_1D(hfov=hfov_samples, wvln=wvln, plane=plane, ray_aiming=ray_aiming)
 
-        # Handle possible NaN values and convert to percentage
-        values = [
-            t.item() * 100 if not math.isnan(t.item()) else 0 for t in distortions
-        ]
+    #     # Handle possible NaN values and convert to percentage
+    #     values = [
+    #         t.item() * 100 if not math.isnan(t.item()) else 0 for t in distortions
+    #     ]
 
-        # Create figure
-        fig, ax = plt.subplots(figsize=(8, 8))
-        ax.set_title(f"{plane} Surface Distortion")
+    #     # Create figure
+    #     fig, ax = plt.subplots(figsize=(8, 8))
+    #     ax.set_title(f"{plane} Surface Distortion")
 
-        # Draw distortion curve
-        ax.plot(values, hfov_samples, linestyle="-", color="g", linewidth=1.5)
+    #     # Draw distortion curve
+    #     ax.plot(values, hfov_samples, linestyle="-", color="g", linewidth=1.5)
 
-        # Draw reference line (vertical line)
-        ax.axvline(x=0, color="k", linestyle="-", linewidth=0.8)
+    #     # Draw reference line (vertical line)
+    #     ax.axvline(x=0, color="k", linestyle="-", linewidth=0.8)
 
-        # Set grid
-        ax.grid(True, color="gray", linestyle="-", linewidth=0.5, alpha=1)
+    #     # Set grid
+    #     ax.grid(True, color="gray", linestyle="-", linewidth=0.5, alpha=1)
 
-        # Dynamically adjust x-axis range
-        value = max(abs(v) for v in values)
-        margin = value * 0.2  # 20% margin
-        x_min, x_max = -max(0.2, value + margin), max(0.2, value + margin)
+    #     # Dynamically adjust x-axis range
+    #     value = max(abs(v) for v in values)
+    #     margin = value * 0.2  # 20% margin
+    #     x_min, x_max = -max(0.2, value + margin), max(0.2, value + margin)
 
-        # Set ticks
-        x_ticks = np.linspace(-value, value, 3)
-        y_ticks = np.linspace(0, hfov, 3)
+    #     # Set ticks
+    #     x_ticks = np.linspace(-value, value, 3)
+    #     y_ticks = np.linspace(0, hfov, 3)
 
-        ax.set_xticks(x_ticks)
-        ax.set_yticks(y_ticks)
+    #     ax.set_xticks(x_ticks)
+    #     ax.set_yticks(y_ticks)
 
-        # Format tick labels
-        x_labels = [f"{x:.1f}%" for x in x_ticks]
-        y_labels = [f"{y:.1f}" for y in y_ticks]
+    #     # Format tick labels
+    #     x_labels = [f"{x:.1f}%" for x in x_ticks]
+    #     y_labels = [f"{y:.1f}" for y in y_ticks]
 
-        ax.set_xticklabels(x_labels)
-        ax.set_yticklabels(y_labels)
+    #     ax.set_xticklabels(x_labels)
+    #     ax.set_yticklabels(y_labels)
 
-        # Set axis labels
-        ax.set_xlabel("Distortion (%)")
-        ax.set_ylabel("Field of View (degrees)")
+    #     # Set axis labels
+    #     ax.set_xlabel("Distortion (%)")
+    #     ax.set_ylabel("Field of View (degrees)")
 
-        # Set axis range
-        ax.set_xlim(x_min, x_max)
-        ax.set_ylim(0, hfov)
-        if filename is None:
-            plt.savefig(
-                f"./{plane}_distortion_infinite_mm.png",
-                bbox_inches="tight",
-                format="png",
-                dpi=300,
-            )
-        else:
-            plt.savefig(
-                f"{filename[:-4]}_{plane}_distortion_infinite_mm.png",
-                bbox_inches="tight",
-                format="png",
-                dpi=300,
-            )
+    #     # Set axis range
+    #     ax.set_xlim(x_min, x_max)
+    #     ax.set_ylim(0, hfov)
+    #     if filename is None:
+    #         plt.savefig(
+    #             f"./{plane}_distortion_infinite_mm.png",
+    #             bbox_inches="tight",
+    #             format="png",
+    #             dpi=300,
+    #         )
+    #     else:
+    #         plt.savefig(
+    #             f"{filename[:-4]}_{plane}_distortion_infinite_mm.png",
+    #             bbox_inches="tight",
+    #             format="png",
+    #             dpi=300,
+    #         )
 
     # ====================================================================================
     # Loss functions and lens design constraints
     # ====================================================================================
-    def init_constraints(self):
-        """Initialize constraints for the lens design."""
-        if self.r_sensor < 12.0:
-            self.is_cellphone = True
+    # def init_constraints(self):
+    #     """Initialize constraints for the lens design."""
+    #     if self.r_sensor < 12.0:
+    #         self.is_cellphone = True
 
-            self.dist_min = 0.1
-            self.dist_max = 0.6  # float("inf")
-            self.thickness_min = 0.3
-            self.thickness_max = 1.5
-            self.flange_min = 0.5
-            self.flange_max = float("inf")
+    #         self.dist_min = 0.1
+    #         self.dist_max = 0.6  # float("inf")
+    #         self.thickness_min = 0.3
+    #         self.thickness_max = 1.5
+    #         self.flange_min = 0.5
+    #         self.flange_max = float("inf")
 
-            self.sag_max = 0.8
-            self.grad_max = 1.0
-            self.grad2_max = 100.0
-        else:
-            self.is_cellphone = False
+    #         self.sag_max = 0.8
+    #         self.grad_max = 1.0
+    #         self.grad2_max = 100.0
+    #     else:
+    #         self.is_cellphone = False
 
-            self.dist_min = 0.1
-            self.dist_max = float("inf")
-            self.thickness_min = 0.3
-            self.thickness_max = float("inf")
-            self.flange_min = 0.5
-            self.flange_max = float("inf")
+    #         self.dist_min = 0.1
+    #         self.dist_max = float("inf")
+    #         self.thickness_min = 0.3
+    #         self.thickness_max = float("inf")
+    #         self.flange_min = 0.5
+    #         self.flange_max = float("inf")
 
-            self.sag_max = 8.0
-            self.grad_max = 1.0
-            self.grad2_max = 100.0
+    #         self.sag_max = 8.0
+    #         self.grad_max = 1.0
+    #         self.grad2_max = 100.0
 
-    def loss_reg(self, w_focus=None):
-        """An empirical regularization loss for lens design. By default we should use weight 0.1 * self.loss_reg() in the total loss."""
-        loss_focus = self.loss_infocus()
+    # def loss_reg(self, w_focus=None):
+    #     """An empirical regularization loss for lens design. By default we should use weight 0.1 * self.loss_reg() in the total loss."""
+    #     loss_focus = self.loss_infocus()
 
-        if self.is_cellphone:
-            loss_intersec = self.loss_self_intersec()
-            loss_surf = self.loss_surface()
-            loss_angle = self.loss_ray_angle()
+    #     if self.is_cellphone:
+    #         loss_intersec = self.loss_self_intersec()
+    #         loss_surf = self.loss_surface()
+    #         loss_angle = self.loss_ray_angle()
 
-            w_focus = 2.0 if w_focus is None else w_focus
-            loss_reg = (
-                w_focus * loss_focus
-                + 1.0 * loss_intersec
-                + 1.0 * loss_surf
-                + 0.1 * loss_angle
-            )
-        else:
-            loss_intersec = self.loss_self_intersec()
-            loss_surf = self.loss_surface()
-            loss_angle = self.loss_ray_angle()
+    #         w_focus = 2.0 if w_focus is None else w_focus
+    #         loss_reg = (
+    #             w_focus * loss_focus
+    #             + 1.0 * loss_intersec
+    #             + 1.0 * loss_surf
+    #             + 0.1 * loss_angle
+    #         )
+    #     else:
+    #         loss_intersec = self.loss_self_intersec()
+    #         loss_surf = self.loss_surface()
+    #         loss_angle = self.loss_ray_angle()
 
-            w_focus = 5.0 if w_focus is None else w_focus
-            loss_reg = (
-                w_focus * loss_focus
-                + 1.0 * loss_intersec
-                + 1.0 * loss_surf
-                + 0.05 * loss_angle
-            )
+    #         w_focus = 5.0 if w_focus is None else w_focus
+    #         loss_reg = (
+    #             w_focus * loss_focus
+    #             + 1.0 * loss_intersec
+    #             + 1.0 * loss_surf
+    #             + 0.05 * loss_angle
+    #         )
 
-        return loss_reg
+    #     return loss_reg
 
-    def loss_infocus(self, bound=0.005):
-        """Sample parallel rays and compute RMS loss on the sensor plane, minimize focus loss.
+    # def loss_infocus(self, bound=0.005):
+    #     """Sample parallel rays and compute RMS loss on the sensor plane, minimize focus loss.
 
-        Args:
-            bound (float, optional): bound of RMS loss. Defaults to 0.005 [mm].
-        """
-        focz = self.d_sensor
-        loss = []
-        for wv in WAVE_RGB:
-            # Ray tracing
-            ray = self.sample_parallel(
-                fov_x=0.0, fov_y=0.0, num_rays=SPP_CALC, wvln=wv, entrance_pupil=True
-            )
-            ray, _ = self.trace(ray)
-            p = ray.project_to(focz)
+    #     Args:
+    #         bound (float, optional): bound of RMS loss. Defaults to 0.005 [mm].
+    #     """
+    #     focz = self.d_sensor
+    #     loss = []
+    #     for wv in WAVE_RGB:
+    #         # Ray tracing
+    #         ray = self.sample_parallel(
+    #             fov_x=0.0, fov_y=0.0, num_rays=SPP_CALC, wvln=wv, entrance_pupil=True
+    #         )
+    #         ray, _ = self.trace(ray)
+    #         p = ray.project_to(focz)
 
-            # Calculate RMS spot size as loss function
-            rms_size = torch.sqrt(
-                torch.sum((p**2 + EPSILON) * ray.ra.unsqueeze(-1))
-                / (torch.sum(ray.ra) + EPSILON)
-            )
-            loss.append(max(rms_size, bound))
+    #         # Calculate RMS spot size as loss function
+    #         rms_size = torch.sqrt(
+    #             torch.sum((p**2 + EPSILON) * ray.ra.unsqueeze(-1))
+    #             / (torch.sum(ray.ra) + EPSILON)
+    #         )
+    #         loss.append(max(rms_size, bound))
 
-        loss_avg = sum(loss) / len(loss)
-        return loss_avg
+    #     loss_avg = sum(loss) / len(loss)
+    #     return loss_avg
 
-    def loss_rms(self, num_grid=GEO_GRID, depth=DEPTH, num_rays=SPP_CALC, importance_sampling=False):
-        """Compute RGB RMS error per pixel, forward rms error.
+    # def loss_rms(self, num_grid=GEO_GRID, depth=DEPTH, num_rays=SPP_CALC, importance_sampling=False):
+    #     """Compute RGB RMS error per pixel, forward rms error.
 
-        Can also revise this function to plot PSF.
-        """
-        # PSF and RMS by patch
-        all_rms_errors = []
-        for i, wvln in enumerate([WAVE_RGB[1], WAVE_RGB[0], WAVE_RGB[2]]):
-            ray = self.sample_point_source(
-                depth=depth,
-                num_rays=num_rays,
-                num_grid=num_grid,
-                wvln=wvln,
-                importance_sampling=importance_sampling,
-            )
-            ray = self.trace2sensor(ray)
+    #     Can also revise this function to plot PSF.
+    #     """
+    #     # PSF and RMS by patch
+    #     all_rms_errors = []
+    #     for i, wvln in enumerate([WAVE_RGB[1], WAVE_RGB[0], WAVE_RGB[2]]):
+    #         ray = self.sample_point_source(
+    #             depth=depth,
+    #             num_rays=num_rays,
+    #             num_grid=num_grid,
+    #             wvln=wvln,
+    #             importance_sampling=importance_sampling,
+    #         )
+    #         ray = self.trace2sensor(ray)
             
-            # Green light point center for reference
-            if i == 0:
-                pointc_green = ((ray.o[..., :2] * ray.ra.unsqueeze(-1)).sum(-2) /
-                                ray.ra.sum(-1).add(EPSILON).unsqueeze(-1))  # shape [num_grid, num_grid, 2]
-                pointc_green = pointc_green.unsqueeze(-2).repeat(
-                    1, 1, num_rays, 1
-                )  # shape [1, num_grid, num_grid, 2]
+    #         # Green light point center for reference
+    #         if i == 0:
+    #             pointc_green = ((ray.o[..., :2] * ray.ra.unsqueeze(-1)).sum(-2) /
+    #                             ray.ra.sum(-1).add(EPSILON).unsqueeze(-1))  # shape [num_grid, num_grid, 2]
+    #             pointc_green = pointc_green.unsqueeze(-2).repeat(
+    #                 1, 1, num_rays, 1
+    #             )  # shape [1, num_grid, num_grid, 2]
 
-            # Calculate RMS error
-            o2_norm = (ray.o[..., :2] - pointc_green) * ray.ra.unsqueeze(-1)
-            o2_norm = o2_norm[num_grid // 2, num_grid // 2, ...]
-            ray.ra = ray.ra[num_grid // 2, num_grid // 2, ...]
+    #         # Calculate RMS error
+    #         o2_norm = (ray.o[..., :2] - pointc_green) * ray.ra.unsqueeze(-1)
+    #         o2_norm = o2_norm[num_grid // 2, num_grid // 2, ...]
+    #         ray.ra = ray.ra[num_grid // 2, num_grid // 2, ...]
 
-            rms_error = torch.mean((((o2_norm ** 2).sum(-1) * ray.ra).sum(-1) /
-                         (ray.ra.sum(-1) + EPSILON)).sqrt()
-                       )
-            all_rms_errors.append(rms_error)
+    #         rms_error = torch.mean((((o2_norm ** 2).sum(-1) * ray.ra).sum(-1) /
+    #                      (ray.ra.sum(-1) + EPSILON)).sqrt()
+    #                    )
+    #         all_rms_errors.append(rms_error)
 
-        avg_rms_error = torch.stack(all_rms_errors).mean(dim=0)
-        return avg_rms_error
+    #     avg_rms_error = torch.stack(all_rms_errors).mean(dim=0)
+    #     return avg_rms_error
 
-    def loss_rms_infinite(self, num_grid=GEO_GRID, depth=DEPTH, num_rays=SPP_CALC):
-        """Compute RGB RMS error per pixel using Zernike polynomials.
+    # def loss_rms_infinite(self, num_grid=GEO_GRID, depth=DEPTH, num_rays=SPP_CALC):
+    #     """Compute RGB RMS error per pixel using Zernike polynomials.
 
-        Args:
-            num_fields: Number of fields. Defaults to 3.
-            depth: object space depth. Defaults to DEPTH.
-        """
-        # calculate fov_x and fov_y
-        [H, W] = self.sensor_res
-        tan_fov_y = np.sqrt(np.tan(self.hfov)**2 / (1 + W**2 / H**2))
-        tan_fov_x = np.sqrt(np.tan(self.hfov)**2 - tan_fov_y**2)
-        fov_y = np.rad2deg(np.arctan(tan_fov_y))
-        fov_x = np.rad2deg(np.arctan(tan_fov_x))
-        fov_y = torch.linspace(0.0, fov_y, num_grid).tolist()
-        fov_x = torch.linspace(0.0, fov_x, num_grid).tolist()
+    #     Args:
+    #         num_fields: Number of fields. Defaults to 3.
+    #         depth: object space depth. Defaults to DEPTH.
+    #     """
+    #     # calculate fov_x and fov_y
+    #     [H, W] = self.sensor_res
+    #     tan_fov_y = np.sqrt(np.tan(self.hfov)**2 / (1 + W**2 / H**2))
+    #     tan_fov_x = np.sqrt(np.tan(self.hfov)**2 - tan_fov_y**2)
+    #     fov_y = np.rad2deg(np.arctan(tan_fov_y))
+    #     fov_x = np.rad2deg(np.arctan(tan_fov_x))
+    #     fov_y = torch.linspace(0.0, fov_y, num_grid).tolist()
+    #     fov_x = torch.linspace(0.0, fov_x, num_grid).tolist()
         
-        # calculate RMS error
-        all_rms_errors = []
-        all_rms_radii = []
-        for i, wvln in enumerate([WAVE_RGB[1], WAVE_RGB[0], WAVE_RGB[2]]):
-            # Ray tracing
-            ray = self.sample_parallel(
-                fov_x=fov_x, fov_y=fov_y, num_rays=num_rays, wvln=wvln, depth=depth
-            )
-            ray = self.trace2sensor(ray)
+    #     # calculate RMS error
+    #     all_rms_errors = []
+    #     all_rms_radii = []
+    #     for i, wvln in enumerate([WAVE_RGB[1], WAVE_RGB[0], WAVE_RGB[2]]):
+    #         # Ray tracing
+    #         ray = self.sample_parallel(
+    #             fov_x=fov_x, fov_y=fov_y, num_rays=num_rays, wvln=wvln, depth=depth
+    #         )
+    #         ray = self.trace2sensor(ray)
 
-            # Green light point center for reference
-            if i == 0:
-                pointc_green = (ray.o[..., :2] * ray.ra.unsqueeze(-1)).sum(
-                    -2
-                ) / ray.ra.sum(-1).add(EPSILON).unsqueeze(-1)   # shape [1, num_fields, 2]
-                pointc_green = pointc_green.unsqueeze(-2).repeat(
-                    1, 1, SPP_PSF, 1
-                )  # shape [num_fields, num_fields, num_rays, 2]
+    #         # Green light point center for reference
+    #         if i == 0:
+    #             pointc_green = (ray.o[..., :2] * ray.ra.unsqueeze(-1)).sum(
+    #                 -2
+    #             ) / ray.ra.sum(-1).add(EPSILON).unsqueeze(-1)   # shape [1, num_fields, 2]
+    #             pointc_green = pointc_green.unsqueeze(-2).repeat(
+    #                 1, 1, SPP_PSF, 1
+    #             )  # shape [num_fields, num_fields, num_rays, 2]
 
-            # Calculate RMS error for different FoVs
-            o2_norm = (ray.o[..., :2] - pointc_green) * ray.ra.unsqueeze(-1)
+    #         # Calculate RMS error for different FoVs
+    #         o2_norm = (ray.o[..., :2] - pointc_green) * ray.ra.unsqueeze(-1)
             
-            # error
-            rms_error = torch.mean(
-                (((o2_norm**2).sum(-1) * ray.ra).sum(-1) / 
-                            (ray.ra.sum(-1) + EPSILON)
-                            ).sqrt()
-            )
+    #         # error
+    #         rms_error = torch.mean(
+    #             (((o2_norm**2).sum(-1) * ray.ra).sum(-1) / 
+    #                         (ray.ra.sum(-1) + EPSILON)
+    #                         ).sqrt()
+    #         )
             
-            # radius
-            rms_radius = torch.mean(
-                ((o2_norm**2).sum(-1) * ray.ra).sqrt().max(dim=-1).values
-            )
-            all_rms_errors.append(rms_error)
-            all_rms_radii.append(rms_radius)
+    #         # radius
+    #         rms_radius = torch.mean(
+    #             ((o2_norm**2).sum(-1) * ray.ra).sqrt().max(dim=-1).values
+    #         )
+    #         all_rms_errors.append(rms_error)
+    #         all_rms_radii.append(rms_radius)
 
-        # Calculate and print average across wavelengths
-        avg_rms_error = torch.stack(all_rms_errors).mean(dim=0)
-        avg_rms_radius = torch.stack(all_rms_radii).mean(dim=0)
+    #     # Calculate and print average across wavelengths
+    #     avg_rms_error = torch.stack(all_rms_errors).mean(dim=0)
+    #     avg_rms_radius = torch.stack(all_rms_radii).mean(dim=0)
         
-        return avg_rms_error
+    #     return avg_rms_error
     
-    def loss_mtf(self, relative_fov=[0.0, 0.7, 1.0], depth=DEPTH, wvln=DEFAULT_WAVE):
-        """Loss function designed on the MTF. We want to maximize MTF values."""
-        loss = 0.0
-        for fov in relative_fov:
-            # ==> Calculate PSF
-            point = torch.tensor([fov, fov, depth])
-            psf = self.psf(points=point, wvln=wvln, ks=256)
+    # def loss_mtf(self, relative_fov=[0.0, 0.7, 1.0], depth=DEPTH, wvln=DEFAULT_WAVE):
+    #     """Loss function designed on the MTF. We want to maximize MTF values."""
+    #     loss = 0.0
+    #     for fov in relative_fov:
+    #         # ==> Calculate PSF
+    #         point = torch.tensor([fov, fov, depth])
+    #         psf = self.psf(points=point, wvln=wvln, ks=256)
 
-            # ==> Calculate MTF
-            x = torch.linspace(-1, 1, psf.shape[1]) * self.pixel_size * psf.shape[1] / 2
-            y = torch.linspace(-1, 1, psf.shape[0]) * self.pixel_size * psf.shape[0] / 2
+    #         # ==> Calculate MTF
+    #         x = torch.linspace(-1, 1, psf.shape[1]) * self.pixel_size * psf.shape[1] / 2
+    #         y = torch.linspace(-1, 1, psf.shape[0]) * self.pixel_size * psf.shape[0] / 2
 
-            # Extract 1D PSFs along the sagittal and tangential directions
-            center_x = psf.shape[1] // 2
-            center_y = psf.shape[0] // 2
-            sagittal_psf = psf[center_y, :]
-            tangential_psf = psf[:, center_x]
+    #         # Extract 1D PSFs along the sagittal and tangential directions
+    #         center_x = psf.shape[1] // 2
+    #         center_y = psf.shape[0] // 2
+    #         sagittal_psf = psf[center_y, :]
+    #         tangential_psf = psf[:, center_x]
 
-            # Fourier Transform to get the MTFs
-            sagittal_mtf = torch.abs(torch.fft.fft(sagittal_psf))
-            tangential_mtf = torch.abs(torch.fft.fft(tangential_psf))
+    #         # Fourier Transform to get the MTFs
+    #         sagittal_mtf = torch.abs(torch.fft.fft(sagittal_psf))
+    #         tangential_mtf = torch.abs(torch.fft.fft(tangential_psf))
 
-            # Normalize the MTFs
-            sagittal_mtf /= sagittal_mtf.max().detach()
-            tangential_mtf /= tangential_mtf.max().detach()
-            delta_x = self.pixel_size
+    #         # Normalize the MTFs
+    #         sagittal_mtf /= sagittal_mtf.max().detach()
+    #         tangential_mtf /= tangential_mtf.max().detach()
+    #         delta_x = self.pixel_size
 
-            # Create frequency axis in cycles/mm
-            freq = np.fft.fftfreq(psf.shape[0], delta_x)
+    #         # Create frequency axis in cycles/mm
+    #         freq = np.fft.fftfreq(psf.shape[0], delta_x)
 
-            # Only keep the positive frequencies
-            positive_freq_idx = freq > 0
+    #         # Only keep the positive frequencies
+    #         positive_freq_idx = freq > 0
 
-            loss += torch.sum(
-                sagittal_mtf[positive_freq_idx] + tangential_mtf[positive_freq_idx]
-            ) / len(positive_freq_idx)
+    #         loss += torch.sum(
+    #             sagittal_mtf[positive_freq_idx] + tangential_mtf[positive_freq_idx]
+    #         ) / len(positive_freq_idx)
 
-        return -loss
+    #     return -loss
 
-    def loss_fov(self, depth=DEPTH):
-        """Trace rays from full FoV and converge them to the edge of the sensor. This loss term can constrain the FoV of the lens."""
-        raise NotImplementedError("Need to check this function.")
-        ray = self.sample_point_source_2D(depth=depth, num_rays=7, entrance_pupil=True)
-        ray = self.trace2sensor(ray)
-        loss = (
-            (ray.o[:, 0] * ray.ra).sum() / (ray.ra.sum() + EPSILON) - self.r_sensor
-        ).abs()
-        return loss
+    # def loss_fov(self, depth=DEPTH):
+    #     """Trace rays from full FoV and converge them to the edge of the sensor. This loss term can constrain the FoV of the lens."""
+    #     raise NotImplementedError("Need to check this function.")
+    #     ray = self.sample_point_source_2D(depth=depth, num_rays=7, entrance_pupil=True)
+    #     ray = self.trace2sensor(ray)
+    #     loss = (
+    #         (ray.o[:, 0] * ray.ra).sum() / (ray.ra.sum() + EPSILON) - self.r_sensor
+    #     ).abs()
+    #     return loss
 
-    def loss_surface(self):
-        """Penalize large sag, first-order derivative, and second-order derivative to prevent surface from being too curved."""
-        sag_max = self.sag_max
-        grad_max = self.grad_max
-        grad2_max = self.grad2_max
+    # def loss_surface(self):
+    #     """Penalize large sag, first-order derivative, and second-order derivative to prevent surface from being too curved."""
+    #     sag_max = self.sag_max
+    #     grad_max = self.grad_max
+    #     grad2_max = self.grad2_max
 
-        loss = 0.0
-        for i in self.find_diff_surf():
-            x_ls = torch.linspace(0.0, 1.0, 20).to(self.device) * self.surfaces[i].r
-            y_ls = torch.zeros_like(x_ls)
+    #     loss = 0.0
+    #     for i in self.find_diff_surf():
+    #         x_ls = torch.linspace(0.0, 1.0, 20).to(self.device) * self.surfaces[i].r
+    #         y_ls = torch.zeros_like(x_ls)
 
-            # Sag
-            sag_ls = self.surfaces[i].sag(x_ls, y_ls)
-            loss += max(sag_ls.max() - sag_ls.min(), sag_max)
+    #         # Sag
+    #         sag_ls = self.surfaces[i].sag(x_ls, y_ls)
+    #         loss += max(sag_ls.max() - sag_ls.min(), sag_max)
 
-            # 1st-order derivative
-            grad_ls = self.surfaces[i].dfdxyz(x_ls, y_ls)[0]
-            loss += 10 * max(grad_ls.abs().max(), grad_max)
+    #         # 1st-order derivative
+    #         grad_ls = self.surfaces[i].dfdxyz(x_ls, y_ls)[0]
+    #         loss += 10 * max(grad_ls.abs().max(), grad_max)
 
-            # 2nd-order derivative
-            grad2_ls = self.surfaces[i].d2fdxyz2(x_ls, y_ls)[0]
-            loss += 10 * max(grad2_ls.abs().max(), grad2_max)
+    #         # 2nd-order derivative
+    #         grad2_ls = self.surfaces[i].d2fdxyz2(x_ls, y_ls)[0]
+    #         loss += 10 * max(grad2_ls.abs().max(), grad2_max)
 
-        return loss
+    #     return loss
 
-    def loss_self_intersec(self):
-        """Loss function to avoid self-intersection. Loss is designed by the distance to the next surfaces."""
-        dist_min = self.dist_min
-        dist_max = self.dist_max
-        thickness_min = self.thickness_min
-        thickness_max = self.thickness_max
-        flange_min = self.flange_min
-        flange_max = self.flange_max
+    # def loss_self_intersec(self):
+    #     """Loss function to avoid self-intersection. Loss is designed by the distance to the next surfaces."""
+    #     dist_min = self.dist_min
+    #     dist_max = self.dist_max
+    #     thickness_min = self.thickness_min
+    #     thickness_max = self.thickness_max
+    #     flange_min = self.flange_min
+    #     flange_max = self.flange_max
 
-        loss_min = 0.0
-        loss_max = 0.0
+    #     loss_min = 0.0
+    #     loss_max = 0.0
 
-        # Constraints for distance/thickness between surfaces
-        for i in range(len(self.surfaces) - 1):
-            current_surf = self.surfaces[i]
-            next_surf = self.surfaces[i + 1]
+    #     # Constraints for distance/thickness between surfaces
+    #     for i in range(len(self.surfaces) - 1):
+    #         current_surf = self.surfaces[i]
+    #         next_surf = self.surfaces[i + 1]
 
-            r = torch.linspace(0.0, 1.0, 20).to(self.device) * current_surf.r
-            z_front = current_surf.surface_with_offset(r, 0)
-            z_next = next_surf.surface_with_offset(r, 0)
+    #         r = torch.linspace(0.0, 1.0, 20).to(self.device) * current_surf.r
+    #         z_front = current_surf.surface_with_offset(r, 0)
+    #         z_next = next_surf.surface_with_offset(r, 0)
 
-            # Minimum distance between surfaces
-            dist_min = torch.min(z_next - z_front)
-            if self.surfaces[i].mat2.name != "air":
-                loss_min += min(thickness_min, dist_min)
-            else:
-                loss_min += min(dist_min, dist_min)
+    #         # Minimum distance between surfaces
+    #         dist_min = torch.min(z_next - z_front)
+    #         if self.surfaces[i].mat2.name != "air":
+    #             loss_min += min(thickness_min, dist_min)
+    #         else:
+    #             loss_min += min(dist_min, dist_min)
 
-            # Maximum distance between surfaces
-            dist_max = torch.max(z_next - z_front)
-            if self.surfaces[i].mat2.name != "air":
-                pass
-            else:
-                loss_max += max(thickness_max, dist_max)
+    #         # Maximum distance between surfaces
+    #         dist_max = torch.max(z_next - z_front)
+    #         if self.surfaces[i].mat2.name != "air":
+    #             pass
+    #         else:
+    #             loss_max += max(thickness_max, dist_max)
 
-        # Constraints for distance to the sensor (flange distance)
-        last_surf = self.surfaces[-1]
-        r = torch.linspace(0.0, 1.0, 20).to(self.device) * last_surf.r
-        z_last_surf = self.d_sensor - last_surf.surface_with_offset(r, 0)
-        loss_min += min(flange_min, torch.min(z_last_surf))
-        loss_max += max(flange_max, torch.min(z_last_surf))
+    #     # Constraints for distance to the sensor (flange distance)
+    #     last_surf = self.surfaces[-1]
+    #     r = torch.linspace(0.0, 1.0, 20).to(self.device) * last_surf.r
+    #     z_last_surf = self.d_sensor - last_surf.surface_with_offset(r, 0)
+    #     loss_min += min(flange_min, torch.min(z_last_surf))
+    #     loss_max += max(flange_max, torch.min(z_last_surf))
 
-        return loss_max - loss_min
+    #     return loss_max - loss_min
 
-    def loss_ray_angle(self, target=0.5, depth=DEPTH):
-        """Loss function designed to penalize large incident angle rays.
+    # def loss_ray_angle(self, target=0.5, depth=DEPTH):
+    #     """Loss function designed to penalize large incident angle rays.
 
-        Reference value: > 0.7
-        """
-        # Sample rays [512, M, M]
-        M = GEO_GRID
-        spp = 512
-        ray = self.sample_point_source(depth=depth, num_rays=spp, num_grid=M)
+    #     Reference value: > 0.7
+    #     """
+    #     # Sample rays [512, M, M]
+    #     M = GEO_GRID
+    #     spp = 512
+    #     ray = self.sample_point_source(depth=depth, num_rays=spp, num_grid=M)
 
-        # Ray tracing
-        ray, _ = self.trace(ray)
+    #     # Ray tracing
+    #     ray, _ = self.trace(ray)
 
-        # Loss (we want to maximize ray angle term)
-        loss = ray.obliq.min()
-        loss = min(loss, target)
+    #     # Loss (we want to maximize ray angle term)
+    #     loss = ray.obliq.min()
+    #     loss = min(loss, target)
 
-        return -loss
+    #     return -loss
 
     # ====================================================================================
     # Optimization
@@ -3196,143 +3196,143 @@ class GeoLens(Lens):
         optimizer = torch.optim.Adam(params)
         return optimizer
 
-    def optimize(
-        self,
-        lrs=[5e-4, 1e-4, 0.1, 1e-3],
-        decay=0.01,
-        iterations=2000,
-        test_per_iter=100,
-        centroid=False,
-        optim_mat=False,
-        match_mat=False,
-        shape_control=True,
-        importance_sampling=False,
-        result_dir="./results",
-    ):
-        """Optimize the lens by minimizing rms errors.
+    # def optimize(
+    #     self,
+    #     lrs=[5e-4, 1e-4, 0.1, 1e-3],
+    #     decay=0.01,
+    #     iterations=2000,
+    #     test_per_iter=100,
+    #     centroid=False,
+    #     optim_mat=False,
+    #     match_mat=False,
+    #     shape_control=True,
+    #     importance_sampling=False,
+    #     result_dir="./results",
+    # ):
+    #     """Optimize the lens by minimizing rms errors.
 
-        Debug hints:
-            *, Slowly and continuously update!
-            1, thickness (fov and ttl should match)
-            2, alpha order (higher is better but more sensitive)
-            3, learning rate and decay (prefer smaller lr and decay)
-            4, correct params range
-        """
-        # Preparation
-        depth = DEPTH
-        num_grid = 31
-        spp = 1024
+    #     Debug hints:
+    #         *, Slowly and continuously update!
+    #         1, thickness (fov and ttl should match)
+    #         2, alpha order (higher is better but more sensitive)
+    #         3, learning rate and decay (prefer smaller lr and decay)
+    #         4, correct params range
+    #     """
+    #     # Preparation
+    #     depth = DEPTH
+    #     num_grid = 31
+    #     spp = 1024
 
-        sample_rays_per_iter = 5 * test_per_iter if centroid else test_per_iter
+    #     sample_rays_per_iter = 5 * test_per_iter if centroid else test_per_iter
 
-        result_dir = (
-            result_dir + "/" + datetime.now().strftime("%m%d-%H%M%S") + "-DesignLens"
-        )
-        os.makedirs(result_dir, exist_ok=True)
-        if not logging.getLogger().hasHandlers():
-            set_logger(result_dir)
-        logging.info(
-            f"lr:{lrs}, decay:{decay}, iterations:{iterations}, spp:{spp}, grid:{num_grid}."
-        )
+    #     result_dir = (
+    #         result_dir + "/" + datetime.now().strftime("%m%d-%H%M%S") + "-DesignLens"
+    #     )
+    #     os.makedirs(result_dir, exist_ok=True)
+    #     if not logging.getLogger().hasHandlers():
+    #         set_logger(result_dir)
+    #     logging.info(
+    #         f"lr:{lrs}, decay:{decay}, iterations:{iterations}, spp:{spp}, grid:{num_grid}."
+    #     )
 
-        optimizer = self.get_optimizer(lrs, decay, optim_mat=optim_mat)
-        scheduler = get_cosine_schedule_with_warmup(
-            optimizer, num_warmup_steps=200, num_training_steps=iterations
-        )
+    #     optimizer = self.get_optimizer(lrs, decay, optim_mat=optim_mat)
+    #     scheduler = get_cosine_schedule_with_warmup(
+    #         optimizer, num_warmup_steps=200, num_training_steps=iterations
+    #     )
 
-        # Training loop
-        pbar = tqdm(total=iterations + 1, desc="Progress", postfix={"loss": 0})
-        for i in range(iterations + 1):
-            # ===> Evaluate the lens
-            if i % test_per_iter == 0:
-                with torch.no_grad():
-                    if i > 0:
-                        if shape_control:
-                            self.correct_shape()
+    #     # Training loop
+    #     pbar = tqdm(total=iterations + 1, desc="Progress", postfix={"loss": 0})
+    #     for i in range(iterations + 1):
+    #         # ===> Evaluate the lens
+    #         if i % test_per_iter == 0:
+    #             with torch.no_grad():
+    #                 if i > 0:
+    #                     if shape_control:
+    #                         self.correct_shape()
 
-                        if optim_mat and match_mat:
-                            self.match_materials()
+    #                     if optim_mat and match_mat:
+    #                         self.match_materials()
 
-                    self.write_lens_json(f"{result_dir}/iter{i}.json")
-                    self.analysis(
-                        f"{result_dir}/iter{i}",
-                        multi_plot=False,
-                    )
+    #                 self.write_lens_json(f"{result_dir}/iter{i}.json")
+    #                 self.analysis(
+    #                     f"{result_dir}/iter{i}",
+    #                     multi_plot=False,
+    #                 )
 
-            # ===> Sample new rays and calculate center
-            if i % sample_rays_per_iter == 0:
-                with torch.no_grad():
-                    # Sample rays
-                    rays_backup = []
-                    for wv in WAVE_RGB:
-                        ray = self.sample_point_source(
-                            depth=depth,
-                            num_rays=spp,
-                            num_grid=num_grid,
-                            wvln=wv,
-                            importance_sampling=importance_sampling,
-                        )
-                        rays_backup.append(ray)
+    #         # ===> Sample new rays and calculate center
+    #         if i % sample_rays_per_iter == 0:
+    #             with torch.no_grad():
+    #                 # Sample rays
+    #                 rays_backup = []
+    #                 for wv in WAVE_RGB:
+    #                     ray = self.sample_point_source(
+    #                         depth=depth,
+    #                         num_rays=spp,
+    #                         num_grid=num_grid,
+    #                         wvln=wv,
+    #                         importance_sampling=importance_sampling,
+    #                     )
+    #                     rays_backup.append(ray)
 
-                    # Calculate ray centers
-                    if centroid:
-                        center_p = -self.psf_center(
-                            point=ray.o[:, :, 0, :], method="chief_ray"
-                        )
-                        center_p = center_p.unsqueeze(-2).repeat(1, 1, spp, 1)
-                    else:
-                        center_p = -self.psf_center(
-                            point=ray.o[:, :, 0, :], method="pinhole"
-                        )
-                        center_p = center_p.unsqueeze(-2).repeat(1, 1, spp, 1)
+    #                 # Calculate ray centers
+    #                 if centroid:
+    #                     center_p = -self.psf_center(
+    #                         point=ray.o[:, :, 0, :], method="chief_ray"
+    #                     )
+    #                     center_p = center_p.unsqueeze(-2).repeat(1, 1, spp, 1)
+    #                 else:
+    #                     center_p = -self.psf_center(
+    #                         point=ray.o[:, :, 0, :], method="pinhole"
+    #                     )
+    #                     center_p = center_p.unsqueeze(-2).repeat(1, 1, spp, 1)
 
-            # ===> Optimize lens by minimizing RMS
-            loss_rms = []
-            for j, wv in enumerate(WAVE_RGB):
-                # Ray tracing
-                ray = rays_backup[j].clone()
-                ray = self.trace2sensor(ray)
-                xy = ray.o[..., :2]  # [h, w, spp, 2]
-                ra = ray.ra.clone().detach()  # [h, w, spp]
-                xy_norm = (xy - center_p) * ra.unsqueeze(-1)  # [h, w, spp, 2]
+    #         # ===> Optimize lens by minimizing RMS
+    #         loss_rms = []
+    #         for j, wv in enumerate(WAVE_RGB):
+    #             # Ray tracing
+    #             ray = rays_backup[j].clone()
+    #             ray = self.trace2sensor(ray)
+    #             xy = ray.o[..., :2]  # [h, w, spp, 2]
+    #             ra = ray.ra.clone().detach()  # [h, w, spp]
+    #             xy_norm = (xy - center_p) * ra.unsqueeze(-1)  # [h, w, spp, 2]
 
-                # Use only quater of the sensor
-                xy_norm = xy_norm[num_grid // 2 :, num_grid // 2 :, :, :]
-                ra = ra[num_grid // 2 :, num_grid // 2 :, :]  # [h/2, w/2, spp]
+    #             # Use only quater of the sensor
+    #             xy_norm = xy_norm[num_grid // 2 :, num_grid // 2 :, :, :]
+    #             ra = ra[num_grid // 2 :, num_grid // 2 :, :]  # [h/2, w/2, spp]
 
-                # Weight mask
-                with torch.no_grad():
-                    weight_mask = (xy_norm.clone().detach() ** 2).sum(-1).sqrt().sum(
-                        -1
-                    ) / (ra.sum(-1) + EPSILON)  # Use L2 error as weight mask
-                    weight_mask /= weight_mask.mean()  # shape of [h/2, w/2]
+    #             # Weight mask
+    #             with torch.no_grad():
+    #                 weight_mask = (xy_norm.clone().detach() ** 2).sum(-1).sqrt().sum(
+    #                     -1
+    #                 ) / (ra.sum(-1) + EPSILON)  # Use L2 error as weight mask
+    #                 weight_mask /= weight_mask.mean()  # shape of [h/2, w/2]
 
-                # Weighted L2 loss
-                # l_rms = torch.mean(xy_norm.abs().sum(-1).sum(-1) / (ra.sum(-1) + EPSILON) * weight_mask)
-                l_rms = torch.mean(
-                    (xy_norm**2 + EPSILON).sum(-1).sqrt().sum(-1)
-                    / (ra.sum(-1) + EPSILON)
-                    * weight_mask
-                )
-                loss_rms.append(l_rms)
+    #             # Weighted L2 loss
+    #             # l_rms = torch.mean(xy_norm.abs().sum(-1).sum(-1) / (ra.sum(-1) + EPSILON) * weight_mask)
+    #             l_rms = torch.mean(
+    #                 (xy_norm**2 + EPSILON).sum(-1).sqrt().sum(-1)
+    #                 / (ra.sum(-1) + EPSILON)
+    #                 * weight_mask
+    #             )
+    #             loss_rms.append(l_rms)
 
-            loss_rms = sum(loss_rms) / len(loss_rms)
+    #         loss_rms = sum(loss_rms) / len(loss_rms)
 
-            # Total loss
-            loss_reg = self.loss_reg()
-            w_reg = 0.1
-            L_total = loss_rms + w_reg * loss_reg
+    #         # Total loss
+    #         loss_reg = self.loss_reg()
+    #         w_reg = 0.1
+    #         L_total = loss_rms + w_reg * loss_reg
 
-            # Back-propagation
-            optimizer.zero_grad()
-            L_total.backward()
-            optimizer.step()
-            scheduler.step()
+    #         # Back-propagation
+    #         optimizer.zero_grad()
+    #         L_total.backward()
+    #         optimizer.step()
+    #         scheduler.step()
 
-            pbar.set_postfix(loss=loss_rms.item())
-            pbar.update(1)
+    #         pbar.set_postfix(loss=loss_rms.item())
+    #         pbar.update(1)
 
-        pbar.close()
+    #     pbar.close()
 
     # ====================================================================================
     # Lens file IO
